pyproject.toml
[tool.poetry]
name = "dct"
version = "2.0.0"
description = "Visualization package for AWS DeepRacer"
authors = ["Jochem Lugtenburg <jochem725>"]

[tool.poetry.dependencies]
python = "^3.7"
opencv-python = "^4.2.0"
ipykernel = "^5.3.0"
numpy = "^1.18.5"
requests = "^2.24.0"
flask = "^1.1.2"
beautifulsoup4 = "^4.9.1"
pysftp = "^0.2.9"
tensorflow = "^2.3.0"
pillow = "^7.2.0"
click = "^7.1.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.8.3"
black = "^20.8b1"

[tool.poetry.scripts]
dct = "dct.cli:cli"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

